Fixed MovieQuizPresenter init to conform MovieQuizViewControllerProtocol
