+++ MovieQuiz/Services/MoviesLoader.swift
@@ -7,10 +7,13 @@
 
 import Foundation
 
+// MARK: - Protocol
 
 protocol MoviesLoading {
     func loadMovies(handler: @escaping (Result<MostPopularMovies, Error>) -> Void)
 }
+
+// MARK: - Structure
 
 struct MovieLoader: MoviesLoading {
     private let networkClient: NetworkRouting
