+++ MovieQuizTests/MoviesLoaderTests.swift
@@ -16,7 +16,6 @@ final class MoviesLoaderTests: XCTestCase {
 		let loader = MovieLoader(networkClient: stubNetworkClient)
 
         //When
-//        let expectation = expectation(description: "Loading expextation") // using this for waiting
 		let expectation = XCTestExpectation(description: "Loading expextation") // for multiple tests better to use XCTestExpectation
 
         loader.loadMovies { result in
@@ -29,7 +28,6 @@ final class MoviesLoaderTests: XCTestCase {
                     XCTFail("Unexpected failure")
             }
         }
-//        waitForExpectations(timeout: 2) // need to wait this time (in secs)
 		wait(for: [expectation], timeout: 2) // use with XCTestExpectation
     }
